gridftp: Add trev gridftp handler
On abnormal terminations, gridftp will asynchronously send a "transfer
event" (trev) to the plugin, which should trigger a shutdown on our end.
Grab that event and set the handle to "done" to cause our plugin to
attempt to shut down properly instead of gridftp coming by and just
sending a kill signal.